libs/python/core/pyproject.toml
[project]
name = "rag-core"
version = "0.1.0"
description = "Shared utilities for the Query Sam Altman workspace"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.1,<7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/rag_core"]
